+++ resources/migrations/013-process_single_facility.down.sql
@@ -1,3 +1,1 @@
 ALTER TABLE facilities_polygons DROP COLUMN starting_node;
-
-DROP FUNCTION IF EXISTS process_facility_isochrones(bigint, varchar, integer, integer, integer);