+++ src/planwise/util/ring.clj
@@ -1,5 +1,6 @@
 (ns planwise.util.ring
-  (:require [ring.util.request :refer [request-url]])
+  (:require [ring.util.request :refer [request-url]]
+            [planwise.model.ident :as ident])
   (:import [java.net URL MalformedURLException]))
 
 (defn- url? [^String s]
@@ -11,3 +12,17 @@
     location
     (let [url (URL. (request-url request))]
       (str (URL. url location)))))
+
+(defn request-ident
+  [request]
+  (:identity request))
+
+(defn request-user-id
+  [request]
+  (-> (request-ident request)
+      ident/user-id))
+
+(defn request-user-email
+  [request]
+  (-> (request-ident request)
+      ident/user-email))